Remove duplicates; Floats to Ints Damage/Health values; Merge DamagePlayer & MobAttackHitbox scripts;Added missing AI files; Remove AI Test scene

+++ Assets/Scripts/Combat/DamageEnemy.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public class DamageEnemy : MonoBehaviour
 {
     [SerializeField]
-    private float attackDamage = 25.0f;
+    private int attackDamage = 25;
 
     private bool isAttacking;
     public bool IsAttacking
